+++ lib/ui/RootPanel.dart
@@ -21,7 +21,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:illinois/main.dart';
-import 'package:illinois/model/Auth2.dart';
+import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/model/Poll.dart';
 import 'package:illinois/service/DeviceCalendar.dart';
 import 'package:illinois/service/ExploreService.dart';