+++ lib/ui/debug/DebugInboxUserInfoPanel.dart
@@ -1,7 +1,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
-import 'package:illinois/service/Auth2.dart';
+import 'package:rokwire_plugin/service/auth2.dart';
 import 'package:illinois/service/Config.dart';
 import 'package:rokwire_plugin/service/log.dart';
 import 'package:rokwire_plugin/service/network.dart';